+++ packages/select/src/components/base-select/Component.tsx
@@ -438,6 +438,7 @@ export const BaseSelect = forwardRef(
                             >
                                 <OptionsList
                                     {...optionsListProps}
+                                    optionsListWidth={optionsListWidth}
                                     flatOptions={flatOptions}
                                     highlightedIndex={highlightedIndex}
                                     open={open}