+++ packages/select/src/components/options-list/Component.tsx
@@ -1,6 +1,8 @@
 import React, { forwardRef, useCallback, useRef } from 'react';
 import cn from 'classnames';
 import mergeRefs from 'react-merge-refs';
+import { Scrollbar } from '@alfalab/core-components-scrollbar';
+import { useMedia } from '@alfalab/hooks';
 import { OptionsListProps, GroupShape, OptionShape } from '../../typings';
 import { Optgroup as DefaultOptgroup } from '../optgroup';
 import { isGroup, useVisibleOptions } from '../../utils';
@@ -30,9 +32,14 @@ export const OptionsList = forwardRef(
             open,
             header,
             footer,
+            optionsListWidth,
+            nativeScrollbar: nativeScrollbarProp,
         }: OptionsListProps,
         ref,
     ) => {
+        let [nativeScrollbar] = useMedia<boolean>([[true, '(max-width: 1023px)']], false);
+        nativeScrollbar = Boolean(nativeScrollbarProp ?? nativeScrollbar);
+
         const renderOption = useCallback(
             (option: OptionShape, index: number) => (
                 <Option key={option.key} {...getOptionProps(option, index)} />
@@ -41,10 +48,16 @@ export const OptionsList = forwardRef(
         );
 
         const listRef = useRef<HTMLDivElement>(null);
+        const scrollbarRef = useRef<HTMLDivElement>(null);
         const counter = createCounter();
         const renderGroup = useCallback(
             (group: GroupShape) => (
-                <Optgroup className={optionGroupClassName} label={group.label} key={group.label} size={size}>
+                <Optgroup
+                    className={optionGroupClassName}
+                    label={group.label}
+                    key={group.label}
+                    size={size}
+                >
                     {group.options.map(option => renderOption(option, counter()))}
                 </Optgroup>
             ),
@@ -52,6 +65,7 @@ export const OptionsList = forwardRef(
         );
 
         useVisibleOptions({
+            ...(!nativeScrollbar && { styleTargetRef: scrollbarRef }),
             visibleOptions,
             listRef,
             open,
@@ -62,18 +76,9 @@ export const OptionsList = forwardRef(
             return null;
         }
 
-        return (
-            <div
-                className={cn(styles.optionsList, styles[size], className)}
-                data-test-id={dataTestId}
-            >
-                {header}
-
-                <div
-                    className={styles.scrollable}
-                    ref={mergeRefs([listRef, ref])}
-                    onScroll={onScroll}
-                >
+        const renderListItems = () => {
+            return (
+                <React.Fragment>
                     {options.map(option =>
                         isGroup(option) ? renderGroup(option) : renderOption(option, counter()),
                     )}
@@ -81,7 +86,50 @@ export const OptionsList = forwardRef(
                     {emptyPlaceholder && options.length === 0 && (
                         <div className={styles.emptyPlaceholder}>{emptyPlaceholder}</div>
                     )}
+                </React.Fragment>
+            );
+        };
+
+        const renderWithCustomScrollbar = () => {
+            const scrollableNodeProps = {
+                onScroll,
+                'data-test-id': dataTestId,
+                ref: ref as React.RefObject<HTMLDivElement>,
+            };
+
+            return (
+                <Scrollbar
+                    className={styles.scrollable}
+                    ref={scrollbarRef}
+                    horizontalAutoStretch={optionsListWidth === 'content'}
+                    scrollableNodeProps={scrollableNodeProps}
+                    contentNodeProps={{ ref: listRef }}
+                >
+                    {renderListItems()}
+                </Scrollbar>
+            );
+        };
+
+        const renderWithNativeScrollbar = () => {
+            return (
+                <div
+                    className={styles.scrollable}
+                    ref={mergeRefs([listRef, ref])}
+                    onScroll={onScroll}
+                >
+                    {renderListItems()}
                 </div>
+            );
+        };
+
+        return (
+            <div
+                {...(nativeScrollbar && { 'data-test-id': dataTestId })}
+                className={cn(styles.optionsList, styles[size], className)}
+            >
+                {header}
+
+                {nativeScrollbar ? renderWithNativeScrollbar() : renderWithCustomScrollbar()}
 
                 {footer}
             </div>