+++ packages/table/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.3.6](https://github.com/core-ds/core-components/compare/@alfalab/core-components-table@1.3.5...@alfalab/core-components-table@1.3.6) (2022-06-08)
+
+**Note:** Version bump only for package @alfalab/core-components-table
+
+
+
+
+
 ## [1.3.5](https://github.com/core-ds/core-components/compare/@alfalab/core-components-table@1.3.4...@alfalab/core-components-table@1.3.5) (2022-06-03)
 
 **Note:** Version bump only for package @alfalab/core-components-table